backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
cryptography
